fix(form-list): patch for rol glitch
There was a glitch where the rol where being added to the form, before their associated bien_raiz was created on the DB. When this happens, and you try to view the detail of this form, it would crash.

The previosuly mentioned error has been fixed and can't happen anymore, but added a patch just in case anyone finds a way to replicate it.