refactor(models): RUT/RUN size
Gave some wiggle room to the size of the RUT/RUN attributte across all the pertinent tables, just in case someone inserts them with '.' in the middle.

In practice, this doesn't really change anything for the use of the application. Rather, it's aimed at possible issues when manually inserting RUT/RUN's through the MySQL console.